successfully connected the database online and merged all backend code into one file. Everything works
